fix: better validation and error messages in server

+++ packages/server/src/routes/event.ts
@@ -22,7 +22,15 @@ export const eventRoutes = (app: Hono) => {
     async (c) => {
       const { userId, rangeStart, rangeEnd } = c.req.valid('query');
       if (userId !== c.var.userId) {
-        throw new HTTPException(400);
+        throw new HTTPException(400, {
+          message: 'userId does not match Authorization header',
+        });
+      }
+
+      if (rangeStart >= rangeEnd) {
+        throw new HTTPException(400, {
+          message: 'rangeEnd must be greater than rangeStart',
+        });
       }
 
       const events = await EventRepo.getAllByUser(userId, [
@@ -54,6 +62,12 @@ export const eventRoutes = (app: Hono) => {
     async (c) => {
       const event = c.req.valid('json');
       const { id } = c.req.valid('param');
+
+      const canEdit = await EventRepo.canEdit(id, c.var.userId);
+      if (!canEdit) {
+        throw new HTTPException(403);
+      }
+
       const res = await EventRepo.replaceOne(id, {
         ...event,
         owner: c.var.userId,
@@ -69,6 +83,12 @@ export const eventRoutes = (app: Hono) => {
     zValidator('param', deleteEventsParamSchema),
     async (c) => {
       const { id } = c.req.valid('param');
+
+      const canEdit = await EventRepo.canEdit(id, c.var.userId);
+      if (!canEdit) {
+        throw new HTTPException(403);
+      }
+
       await EventRepo.deleteOne(id);
 
       return c.body(null, 204);
